Transit+File transfer API, first take
This nowhere near perfect and I'm uncertain about a lot of design decisions, but I think putting all those methods into dedicated modules is a good first step towards a usable API.
